Revision D7.6 (Treadmill | pushThreaded | Saves )
Created Class:
- WorldTreadmill

Changed Classes:
- WorldBlock
- WorldChunk
- WorldSlice
- WorldInstance

- WorldBlock, WorldChunk, WorldSlice, and WorldInstance have proper
constructors that accept their own classtype for copy construction.

- WorldChunk, WorldSlice, and WorldInstance have new methods
computeOffsets and recomputePosition that act like populate but set
position directly rather than by creating a new object.

- WorldInstance has a save method that has an output that follows the
format of "res/save/WORLDNAME/WorldInstance_Generation_X_Y.ser".
Generation is a number that separates chunks of WorldInstances (2000 or
so) so the treadmill can reset and not draw positions that are out of
bounds to a float's maximum data type.

- WorldTreadmill properly pushes new left, right, up, or down instances
in a separate, non-blocking thread. The method also accounts for pushing
arrays of arbitrary lengths, rather than a fixed length of 2x2.